ganti integer jadi double di jumlah dan saldo

+++ database/migrations/2024_03_12_092023_create_transaksi_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('id_transaksi');
             $table->integer('id_kode_perkiraan');
-            $table->bigInteger('jumlah');
+            $table->double('jumlah');
             $table->char('jenis', 1);
             $table->timestamps();
         });
